standalone/astro-apply/pyproject.toml
[tool.poetry]
name = "astro-apply"
version = "0.2.1"
description = "Like `kubectl apply` - but for Astronomer Cloud."
authors = ["solutions <solutions@astronomer.io>"]
readme = "README.md"
repository = "https://github.com/astronomer/astro-apply"
homepage = "https://github.com/astronomer/astro-apply"

[tool.poetry.scripts]
# Entry points for the package https://python-poetry.org/docs/pyproject/#scripts
"astro-apply" = "astro_apply.__main__:cli"

[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^6.0"
gql = {extras = ["requests"], version = "^3.1.0"}
requests = "^2.27.1"
click = "^8.1.2"
python-dotenv = "^0.20.0"
deepmerge = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-mock = "^3.7.0"
isort = "^5.10.1"
black = "^22.3.0"
pre-commit = "^2.18.1"
pyupgrade = "^2.31.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
# https://github.com/psf/black
target-version = ["py37"]
line-length = 120
color = true

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | env
    | venv
    | tests/
)/
'''

[tool.isort]
# https://github.com/timothycrosley/isort/
py_version = 38
line_length = 120

known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true

[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# Directories that are not visited by pytest collector:
norecursedirs =["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]
doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]

# Extra options:
addopts = [
  "--strict-markers",
  "--tb=short",
  "--disable-warnings",
  "--doctest-modules",
  "--doctest-continue-on-failure",
]
